requirements.txt
Django==2.0.3
djangorestframework==3.9.0
numpy==1.15.4
pytz==2018.7
